improved averaging system
The old averaging system had two problems which involved the "need to beat" statistic. First, if you need a 15.525 to beat your PB average, then sometimes a "15.52" would beat the average while other times it wouldn't. This is because the averaging system did not truncate times.

The other problem is that Need To Beat sometimes failed to work properly. It would look for a way to get a full millisecond below the current average, while sometimes a sub-millisecond difference is enough. This is now fixed as well, using a new feature of average.js called integralValueForAverageBelow.